# Measure startup perf as a part of every PR