--- a/package.json
+++ b/package.json
@@ -22,7 +22,6 @@
   },
   "homepage": "https://github.com/paulcbetts/trickline#readme",
   "dependencies": {
-    "@paulcbetts/v8-profiler": "5.6.6",
     "@types/chai": "^3.4.34",
     "@types/debug": "0.0.29",
     "@types/electron": "^1.4.31",