This PR adds measurement of startup performance via Chrome Content Tracing as a part of every PR. At the bottom of every build log, you'll see:

```
load [Load]:
  start:                                0.00ms
  end:                                  2890.11ms
  duration:                             2890.11ms
  parseHTML:                            1.52ms
  javaScript:                           817.45ms
  javaScriptCompile:                    0.41ms
  styles:                               6.45ms
  updateLayerTree:                      1.65ms
  layout:                               39.78ms
  paint:                                2.03ms
  raster:                               21.38ms
  composite:                            0.57ms
  extendedInfo:
    domContentLoaded:                   1089.42ms
    loadTime:                           1086.80ms
    firstPaint:                         0.00ms
    javaScript:
```

This info is also saved in the build artifact as `bigrig.json` (`bigrig` is the tool used to generate this analysis).

## TODO:

- [x] Make sure this works in CI
- [ ] Make sure this works in Docker with .env file