@mnquintana 2017-02-28
What would you think about taking multiple samples per PR? Sample size of 1 would have a pretty wide margin of error 🤔