@paulcbetts 2017-02-28
@mnquintana Not a bad idea, especially since this requires a users.counts call, so network latency affects it.